+++ Editor/VectorSpriteRendererEditor.cs
@@ -21,6 +21,7 @@ namespace VMG.EditorTools
         private SerializedProperty m_Depth;
         private SerializedProperty m_RoundCorners;
         private SerializedProperty m_Trim;
+        private SerializedProperty m_Wiggle;
         private SerializedProperty m_Tint;
         private SerializedProperty m_Material;
         private SerializedProperty m_Texture;
@@ -37,6 +38,7 @@ namespace VMG.EditorTools
             m_Depth = serializedObject.FindProperty("Depth");
             m_RoundCorners = serializedObject.FindProperty("RoundCorners");
             m_Trim = serializedObject.FindProperty("Trim");
+            m_Wiggle = serializedObject.FindProperty("Wiggle");
             m_Tint = serializedObject.FindProperty("Tint");
             m_Material = serializedObject.FindProperty("Material");
             m_Texture = serializedObject.FindProperty("Texture");
@@ -70,6 +72,7 @@ namespace VMG.EditorTools
                 }
                 EditorGUILayout.PropertyField(m_RoundCorners, true);
                 EditorGUILayout.PropertyField(m_Trim, true);
+                EditorGUILayout.PropertyField(m_Wiggle, true);
             }
             if (hasSvg)
             {